requirements.txt
streamlit==1.31.1
Pillow
numpy 